Finalizing FileList Component
* applied Popover over Licenses data
* refactored Licenses and Copyrights renderer
