+++ src/components/FileList/CopyrightsRenderer/CopyrightsRenderer.css
