+++ src/components/FileList/CopyrightsRenderer/index.js
@@ -1,0 +1,1 @@
+export { default } from './CopyrightsRenderer'
