+++ src/components/FileList/LicensesRenderer/LicensesRenderer.css
@@ -1,0 +1,3 @@
+.copyrightPopover{
+  max-height: 200px;
+}
