+++ src/components/FileList/PopoverRenderer.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import isArray from 'lodash/isArray'
+import { Popover } from 'react-bootstrap'
+
+const PopoverComponent = (props) => {
+  return (<Popover id="popover-positioned-left" title={props.title} {...props}>
+    <div style={{ maxHeight: '200px', overflowY: 'auto', overflowX: 'hidden', width: '100%' }}>
+      {props.values && isArray(props.value) ? props.values.map(item => <p key={item}>{item}</p>) : <p>{props.values}</p>}</div>
+  </Popover>)
+}
+
+export default PopoverComponent;
